code refactor for navigation
