+++ src/navigation/StackNav/index.tsx
@@ -1,0 +1,4 @@
+import HomeStack from './HomeStack';
+import SettingsStack from './SettingsStack';
+
+export {HomeStack, SettingsStack};
